gpt-4-output/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.6.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
boto3==1.26.138
botocore==1.29.138
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
cryptography==40.0.2
dataclasses-json==0.5.7
defusedxml==0.7.1
Deprecated==1.2.13
dnspython==2.3.0
et-xmlfile==1.1.0
fastapi==0.95.2
frozenlist==1.3.3
googleapis-common-protos==1.56.4
greenlet==2.0.2
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.53.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
jmespath==1.0.1
joblib==1.2.0
langchain==0.0.137
linkify-it-py==2.0.0
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
monotonic==1.6
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pinecone-client==2.2.1
protobuf==4.22.3
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.7.0
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
social-auth-core==4.4.2
soupsieve==2.4.1
SQLAlchemy==1.4.47
starlette==0.27.0
temporalio==1.1.0
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
types-protobuf==4.22.0.2
typing-inspect==0.8.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
unstructured==0.5.12
urllib3==1.26.16
uvicorn==0.22.0
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.8.2
